Change functions to promise
