+++ src/api.js
@@ -2,8 +2,6 @@ const fs = require('fs');
 const path = require('path');
 const chalk = require('chalk');
 
-const validateOptions = (options) => options === '--validate' || options === '--stats';
-
 const message = (text, color) => chalk.keyword(color)(text);
 
 const formatPath = (route) => route.replace(/\\/g, '/');
@@ -26,8 +24,34 @@ const filterMd = (files) => files.filter((file) => mdExt(file));
 
 const readMd = (route) => fs.readFileSync(route, 'utf-8');
 
-const getLinks = (route) => {
+const readFile = (pathFile) => new Promise((resolve, reject) => {
+  fs.readFile(pathFile, (error, data) => {
+    if (error) {
+      reject(error);
+    }
+    resolve(data);
+  });
+});
+
+const getLinks = (route) => new Promise((resolve, reject) => {
   const links = [];
+  readFile(route)
+    .then((data) => {
+      const regex = /\[(.+?)\]\((https?:\/\/[^\s)]+)\)/g;
+      let match = regex.exec(data);
+      while (match !== null) {
+        links.push({
+          href: match[2],
+          text: match[1],
+          file: route,
+        });
+        match = regex.exec(data);
+      }
+      resolve(links);
+    })
+    .catch((error) => reject(error));
+});
+  /*
   const mdFile = readMd(route);
   const regex = /\[(.+?)\]\((https?:\/\/[^\s)]+)\)/g;
   let match = regex.exec(mdFile);
@@ -40,7 +64,7 @@ const getLinks = (route) => {
     match = regex.exec(mdFile);
   }
   return links;
-};
+  */
 // const readDirectory = (route) => fs.readdirSync(route);
 
 module.exports = {
